Agregado el modal del delete

+++ integradorAngular/src/app/components/products/product-list/product-list.component.ts
@@ -1,7 +1,7 @@
 import { Component, OnInit } from '@angular/core';
+import { NgbModal } from '@ng-bootstrap/ng-bootstrap';
+import { DeleteConfirmationModalComponent } from 'src/app/modals/delete-confirmation-modal/delete-confirmation-modal.component';
 import { ProductsService } from 'src/app/services/products.service';
-
-
 
 @Component({
   selector: 'app-product-list',
@@ -11,13 +11,15 @@ import { ProductsService } from 'src/app/services/products.service';
 export class ProductListComponent implements OnInit {
   productsData: any[] = [];
 
-  constructor(private productService: ProductsService) {}
+  constructor(
+    private productService: ProductsService,
+    private modalService: NgbModal
+  ) {}
 
   ngOnInit(): void {
     // this.productsData = this.productService.getHardcodedProducts();
     this.listProducts();
   }
-
 
   listProducts() {
     const res = this.productService.getProduct();
@@ -26,25 +28,32 @@ export class ProductListComponent implements OnInit {
 
     if (res !== null) {
       this.productsData = res;
-  
-  }else{
-    console.error('La respuesta es nula');
-  }
-}
-
-
-deleteProduct(id: string){
-
-  const res = this.productService.deleteProduct(id); //Acá traigo toda la lista actualizada con el eliminado y la meto en la variable res
-
-  console.log('respuesta del get', res);
-
- if (res !== null && res !== undefined) {
-    this.productsData = res;
-    }else{
-    console.error('La respuesta es nula');
+    } else {
+      console.error('La respuesta es nula');
+    }
   }
 
+  deleteProduct(id: string) {
+    const modalRef = this.modalService.open(DeleteConfirmationModalComponent);
 
-}
+    modalRef.componentInstance.message =
+      '¿Está seguro de que desea eliminar este producto?';
+
+    modalRef.result.then(
+      (result) => {
+        if (result === 'confirm') {
+          const res = this.productService.deleteProduct(id); //Acá traigo toda la lista actualizada con el eliminado y la meto en la variable res
+
+          if (res !== null && res !== undefined) {
+            this.productsData = res;
+          } else {
+            console.error('La respuesta es nula');
+          }
+        }
+      },
+      (reason) => {
+        console.log('Modal de confirmación cerrado sin confirmar: ', reason);
+      }
+    );
+  }
 }
